# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/database_files": {
      "post": {
        "tags": [
          "database_manager"
        ],
        "summary": "Upload Database",
        "operationId": "upload_database_api_v1_database_files_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_database_api_v1_database_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Database Api V1 Database Files Post"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/database_files/{dbname}": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Download Database Endpoint",
        "operationId": "download_database_endpoint_api_v1_database_files__dbname__get",
        "parameters": [
          {
            "name": "dbname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "database_manager"
        ],
        "summary": "Delete Database Endpoint",
        "operationId": "delete_database_endpoint_api_v1_database_files__dbname__delete",
        "parameters": [
          {
            "name": "dbname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Database Endpoint Api V1 Database Files  Dbname  Delete"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get-cell-ids": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Get Cell Ids Endpoint",
        "operationId": "get_cell_ids_endpoint_api_v1_get_cell_ids_get",
        "parameters": [
          {
            "name": "dbname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Cell Ids Endpoint Api V1 Get Cell Ids Get"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get-cell-ids-by-label": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Get Cell Ids By Label Endpoint",
        "operationId": "get_cell_ids_by_label_endpoint_api_v1_get_cell_ids_by_label_get",
        "parameters": [
          {
            "name": "dbname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Cell Ids By Label Endpoint Api V1 Get Cell Ids By Label Get"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get-manual-labels": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Get Manual Labels Endpoint",
        "operationId": "get_manual_labels_endpoint_api_v1_get_manual_labels_get",
        "parameters": [
          {
            "name": "dbname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Manual Labels Endpoint Api V1 Get Manual Labels Get"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get-cell-contour": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Get Cell Contour Endpoint",
        "operationId": "get_cell_contour_endpoint_api_v1_get_cell_contour_get",
        "parameters": [
          {
            "name": "dbname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          },
          {
            "name": "cell_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Cell Contour Endpoint Api V1 Get Cell Contour Get"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get-cell-label": {
      "get": {
        "tags": [
          "database_manager"
        ],
        "summary": "Get Cell Label Endpoint",
        "operationId": "get_cell_label_endpoint_api_v1_get_cell_label_get",
        "parameters": [
          {
            "name": "dbname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dbname"
            }
          },
          {
            "name": "cell_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Cell Label Endpoint Api V1 Get Cell Label Get"
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_database_api_v1_database_files_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_database_api_v1_database_files_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}